backend-django/requirements.txt
arrow==1.2.2
asgiref==3.5.2
astroid==2.11.6
blessed==1.19.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.5.1
dj-rest-auth==2.2.4
Django==4.0.5
django-allauth==0.50.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-import-export==2.8.0
django-picklefield==3.1
django-q==1.3.9
django-sendgrid-v5==1.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
MarkupPy==1.14
mccabe==0.7.0
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pylint==2.14.4
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-dateutil==2.8.2
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
redis==3.5.3
requests==2.28.0
requests-oauthlib==1.3.1
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
tablib==3.2.1
tomli==2.0.1
tomlkit==0.11.0
urllib3==1.26.9
wcwidth==0.2.5
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
